prototype: Isolate all key related operation in panel class. Only work info sample
